% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{classifier_split}
\alias{classifier_split}
\title{split the node for classification}
\usage{
classifier_split(
  X,
  y,
  idxs,
  tree_num = 0,
  sample_weight = NULL,
  max_features = NULL
)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{idxs}{the index of samples used for the split(array of bool values. If TRUE, it is used for the split)}

\item{tree_num}{initial tree_num (default=0)}

\item{sample_weight}{allocate weight for each individual sampe, array-like of shape n (default=None). If None, then samples are equally weighted.}

\item{max_features}{the number of features to consider when looking for the best split}
}
\value{
call: the input setting

node_split: the design matrix for beta after variable splitting
node_split$idxs : the index of samples in the original node
node_split$value: the value for the split node
node_split$impurity: the impurity in the original node
node_split$tree_num: the tree number

node_left: the design matrix for gamma after variable splitting
node_split$idxs : the index of samples in the left node
node_split$value: the value for the left node after the split
node_split$impurity: the impurity in the left node
node_split$tree_num: the tree number which the node belongs to

node_right: the response vector after variable splitting.
node_split$idxs : the index of samples in the right node
node_split$value: the value for the right node after the split
node_split$impurity: the impurity in the right node
node_split$tree_num: the tree number which the node belongs to

n: the number of samples in the original node

feature: the feature selected to do the split

threshold: the threshold to make the split

impurity_reduction: the impurity decreased by the split if the decrease exists
}
\description{
make the split based on figs classification rule
}
