% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{figs.classifier}
\alias{figs.classifier}
\title{figs fit for classification}
\usage{
figs.classifier(
  X,
  y,
  max_rules = 12,
  feature_names = NULL,
  sample_weight = NULL,
  min_impurity_decrease = 0,
  verbose = TRUE,
  max_features = NULL
)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{max_rules}{max total number of rules across all trees}

\item{feature_names}{the name of features in the data}

\item{sample_weight}{allocate weight for each individual sampe, array-like of shape n (default=None). If None, then samples are equally weighted.}

\item{min_impurity_decrease}{A node will be split if this split induces a decrease of the impurity greater than or equal to this value.}

\item{verbose}{bool value, if TRUE print the split feature and threshold}

\item{max_features}{the number of features to consider when looking for the best split}
}
\value{
call: the input setting

node_split: the design matrix for beta after variable splitting
node_split$idxs : the index of samples in the original node
node_split$value: the value for the split node
node_split$impurity: the impurity in the original node
node_split$tree_num: the tree number

node_left: the design matrix for gamma after variable splitting
node_split$idxs : the index of samples in the left node
node_split$value: the value for the left node after the split
node_split$impurity: the impurity in the left node
node_split$tree_num: the tree number which the node belongs to

node_right: the response vector after variable splitting.
node_split$idxs : the index of samples in the right node
node_split$value: the value for the right node after the split
node_split$impurity: the impurity in the right node
node_split$tree_num: the tree number which the node belongs to

trees: tree struction based on the result

n: the number of samples in the original node

feature: the feature selected to do the split

threshold: the threshold to make the split

impurity_reduction: the impurity decreased by the split if the decrease exists
}
\description{
fitting function for fast interpretable greedy-tree sums on classification
}
